EXTRACT/main to makefile
