+++ pracs/01-a_tutorial_introduction/uncomment/Makefile
@@ -1,16 +1,21 @@
 CC				= gcc
 CFLAGS			= -Wall -Werror -Wextra
-NAME			= c-uncomment
-OBJECTS			=
+NAME			= uncomment.a
+OBJECTS			= srcs/strcomment.o \
+				srcs/del_content.o
+PROGRAM			= uncomment
 
 .DEFAULT_GOAL	:= all
 
 .PHONY: all
-all: $(NAME) $(PROGRAMS)
+all: $(NAME) $(PROGRAM)
 
-$(NAME): $(OBJECTS)
+$(NAME): $(OBJECTS) 
 	ar rcs $@ $^
 	ranlib $@
+
+$(PROGRAM):
+	$(CC) $(CFLAGS) -Iincludes srcs/uncomment.c $(NAME) -o $@
 
 .PHONY: clean
 clean:
@@ -21,8 +26,13 @@ clean:
 fclean:
 	echo "DO fclean"
 	/bin/rm $(NAME)
-	/bin/rm $(PROGRAMS)
+	/bin/rm uncomment
 
 .PHONY: re
 re:	clean fclean all
 
+srcs/strcomment.o:
+	$(CC) $(CFLAGS) -c srcs/strcomment.c -o srcs/strcomment.o
+
+srcs/del_content.o:
+	$(CC) $(CFLAGS) -c srcs/del_content.c -o srcs/del_content.o
