+++ pracs/01-a_tutorial_introduction/uncomment/includes/uncomment.h
@@ -3,8 +3,8 @@
 #include <fcntl.h>
 #include <sys/stat.h>
 #include <unistd.h>
-#include <stdlib.h>
-#include <string.h>
+#include <assert.h>
 #define	MAXLEN	4096
-
+const char	*strcomment(const char *haystack, const char *needle);
+size_t		del_content(char *start, char *end);
 #endif
