+++ pracs/01-a_tutorial_introduction/uncomment/srcs/del_content.c
@@ -1,0 +1,12 @@
+#include <stddef.h>
+
+size_t	del_content(char *start, char *end)
+{
+	size_t	ret;
+
+	ret = 0;
+	while ((start + ret) < end)
+		*(start + ret++) = '\0';
+	return (ret);
+}
+
