+++ pracs/01-a_tutorial_introduction/uncomment/srcs/strcomment.c
@@ -1,0 +1,34 @@
+#include <string.h>
+
+const char	*strcomment(const char *haystack, const char *needle)
+{
+	const char	*src_pt;
+	size_t		len;
+	static char qute = 0;
+
+	len = strlen(needle);
+	src_pt = haystack;
+	while (*src_pt) {
+		if (qute) {
+			while (*src_pt != '"' && *src_pt)
+				src_pt++;
+
+			if (*src_pt == '"')
+				qute = ~qute;
+
+		} else {
+			while (*(src_pt) && (*src_pt != *needle) && (*src_pt != '"'))
+				src_pt++;
+
+			if (*src_pt == *needle) {
+				if(!strncmp(src_pt, needle, len))
+					return (src_pt);
+				src_pt++;
+			} else if (*src_pt == '"') {
+				qute = ~qute;
+				src_pt++;
+			}
+		}
+	}
+	return (NULL);
+}
